+++ client/acp/stores/SearchUsersStore.js
@@ -1,7 +1,7 @@
 var Actions       = require('../actions/Actions'),
     AppDispatcher = require('../dispatcher/AppDispatcher'),
     EventEmitter  = require('events').EventEmitter,
-    assign        = require('react/lib/Object.assign'),
+    assign        = require('object-assign'),
     Constants     = require('../Constants'),
     socket        = require('socket'),