+++ client/acp/stores/UsersStore.js
@@ -1,6 +1,6 @@
 var AppDispatcher = require('../dispatcher/AppDispatcher'),
     EventEmitter  = require('events').EventEmitter,
-    assign        = require('react/lib/Object.assign'),
+    assign        = require('object-assign'),
     Constants     = require('../Constants'),
     socket        = require('socket'),